+++ src/game/nlcore/elements/displayable/puppet.ts
@@ -460,6 +460,7 @@ export class Puppet
 
     /**@internal */
     override reset() {
+        super.reset();
         this.state = this.getInitialState();
         this.transformState = this.getInitialTransformState(this.userConfig);
     }